backend/requirements.txt.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
python-jose[cryptography]
